test/src/firebasenotifications/serializer.py
from rest_framework import serializers
from .models import DeviceToken


class DeviceTokenSerializer(serializers.ModelSerializer):
    class Meta:
        model = DeviceToken
        fields = ('userdt','user_id','device_token','created')