test/src/requirements.txt
amqp==2.3.2
anyjson==0.3.3
appdirs==1.4.0
billiard==3.5.0.3
bleach==2.1.3
boto==2.47.0
boto3==1.9.65
botocore==1.12.65
braintree==3.35.0
celery==4.2.0
cffi==1.11.5
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.30.0
django-celery==3.2.1
django_celery_beat==1.0.1
django_celery_results==1.0.1
django-cleanup==1.0.1
django-crispy-forms==1.6.1
django-extensions==2.0.7
django-markdown-deux==1.0.5
django-pagedown==0.1.3
django-recaptcha==1.3.1
django-ses==0.8.2
django-storages==1.5.2
django-summernote==0.8.8.6
djangorestframework==3.9.0
eventlet==0.23.0
greenlet==0.4.13
html5lib==1.0.1
kombu==4.2.1
markdown2==2.3.4
numpy==1.12.1
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pandas==0.19.2
Pillow==4.0.0
psycopg2==2.6.2
pycparser==2.18
pycurl==7.43.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
stripe==1.65.0
vine==1.1.4
webencodings==0.5.1
